+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * Make mod list clip long mod names instead of scrolling horizontally
 * Add UI to switch between play, out-jar, and package
 * Make mod description box scrollable for long text
+* Make Steam icon on mod open the mod's workshop page when clicked
 
 #### v3.30.3 ####
 * Fix package information being lost for patched classes (Alchyr)